Refactor content-tree service
